+++ src/dbt_score/__init__.py
@@ -1,15 +1,16 @@
 """Init dbt_score package."""
 
-from dbt_score.model_filter import ModelFilter, model_filter
-from dbt_score.models import Model
+from dbt_score.models import Model, Source
 from dbt_score.rule import Rule, RuleViolation, Severity, rule
+from dbt_score.rule_filter import RuleFilter, rule_filter
 
 __all__ = [
     "Model",
-    "ModelFilter",
+    "Source",
+    "RuleFilter",
     "Rule",
     "RuleViolation",
     "Severity",
-    "model_filter",
+    "rule_filter",
     "rule",
 ]