+++ src/dbt_score/config.py
@@ -56,7 +56,7 @@ class Config:
         "disabled_rules",
         "inject_cwd_in_python_path",
         "fail_project_under",
-        "fail_any_model_under",
+        "fail_any_item_under",
         "show",
     ]
     _rules_section: Final[str] = "rules"
@@ -71,7 +71,7 @@ class Config:
         self.config_file: Path | None = None
         self.badge_config: BadgeConfig = BadgeConfig()
         self.fail_project_under: float = 5.0
-        self.fail_any_model_under: float = 5.0
+        self.fail_any_item_under: float = 5.0
         self.show: str = "failing-rules"
 
     def set_option(self, option: str, value: Any) -> None: