+++ tests/rules/example.py
@@ -1,6 +1,6 @@
 """Example rules."""
 
-from dbt_score import Model, RuleViolation, model_filter, rule
+from dbt_score import Model, RuleViolation, rule, rule_filter
 
 
 @rule()
@@ -8,7 +8,7 @@ def rule_test_example(model: Model) -> RuleViolation | None:
     """An example rule."""
 
 
-@model_filter
+@rule_filter
 def skip_model1(model: Model) -> bool:
     """An example filter."""
     return model.name != "model1"